kernel-sru-check: get SRU cycle tag and sort signoffs
Sometimes the kernel team needs to respin a kernel to fix a regression
from the last cycle, and the signoff for that kernel will get lost in
the noise of the signoffs for the current cycle.

To address this, pull out the tag that contains the SRU cycle
(identified by the date the cycle *starts*), and then sort the found bug
tasks by the kernel cycles and the major kernel versions.

Signed-off-by: Steve Beattie <steve.beattie@canonical.com>