kernel-abi-check, kernel_lib: add support for checking in esm ppas
Add support for annotating a kernel with which ppa it is published to,
if it is no longer published in the ubuntu archive. Modify the
kernel-abi-script to look up the kernel and meta info from that ppa, to
see if there's a kernel newer than the last USN for that kernel and to
ensure that the ABIs are consistent.

Note that the use of 'release' versus 'release/esm' is inconsistent,
because it is all somewhat inconsistent across UCT.

Signed-off-by: Steve Beattie <steve.beattie@canonical.com>