sis-generate-usn: fix description defaulting to None
cve_lib.lookup_package_override_description() returns None when it
can't find a package in the description table; however the code that
would use the description from the changes files was checking if the
string was "", which in python is not equal to None. Instead, use
the description from the changes files if the description returned
by lookup_package_override_description() is either the empty string
or None.

Signed-off-by: Steve Beattie <steve.beattie@canonical.com>