sis-changes, sis-generate-usn: expand binaries.json format
Convert the binaries.json format to use a dict at the lowest level, so
that we can expand its contents. The current per-binary-dict includes
three fields:

  version -> the actual launchpad binary version
  pocket -> computed from the computed URL of binary package
  origin -> the actual launchpad archive(s) that sis-changes grabbed
    information from.

Alas, the origin field works best for ESM packages, as sis-changes will
typically get run against either the esm-{infra,apps}-security-staging
ppas or in the published esm-{infra,apps}-security ppas themselves. For
things in the archive, these are often published from a variety of
ppas, so heuristics there will be needed (perhaps the URLs are the best
way for that).

Another thing tripped over in testing which is fixed in this commit is
that the changes files for some packages will include binary package
name/description pairs that aren't actually built! Since we have an
accurate (at the time of publication) record of the actual published
binaries, sis-generate-usn can skip recording those entries in the
generated usn.sh.

Also included is some expanded comments to explain the expanded
json structure.

(I think the ideal way forward is that we end up avoiding all the
changes files, and capture all the information we need from launchpad
in the json interchange files, which is why I started to flesh that out
here.)

Signed-off-by: Steve Beattie <steve.beattie@canonical.com>