sis-changes: fix finding pocket for arch specific packages
Some packages only publish binaries for a given arch, for example, on
trusty/esm, where the arch_all arch is i386:

- the linux-signed-lts-xenial package only publishes binaries for amd64
- the linux package publishes some binaries for powerpc and ppc64le only.

Picking just a couple of arches doesn't get them all, and will cause a
key error if neither of the two arches are the ones attempted.

Instead, iterate over all the architectures to try to find the given
binary package/url to try to identify the pocket from.

(The cve_lib interfaces are less friendly here, the
official_architectures field contains 'source' and 'all' as well as the
binary architectures. Would be nice if just had a list of only binary
architectures to compare against.)

That said, it's a little silly to query for the urls when we can
record exactly which ppa we looked up the location of the source and
binaries from, though that will require heuristics when it comes to
the regular archive viz which ppas a given package gets prepared in.

Signed-off-by: Steve Beattie <steve.beattie@canonical.com>