kernel/CVE-2021-23133: update break-fix commits
Original fix was found to cause a regression, leading to a potential
deadlock. It was reverted and a different fix was applied.

Use the new fix for fixing the original vulnerability, but move the old fix
as a break and add its revert as a fix, so we make sure we don't have the
regression on our kernels.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>