usn_lib.USNdb: lookup pkg info if not in cached info
If the packages dict does not have cached info about the passed package
in get_usns() then don't just return no info, but attempt to lookup the
info in the usn db.

The reason to do this is to make the USNdb class less difficult to
work with, so that you don't have to know all the packages you want
to gather at class instantiation time to use the class, letting the
a USNdb instance be re-used in multiple package queries.

Ideally, going forward this class becomes a more generic class for
interacting with the primary USN db as well as subsets of the USN db.

Signed-off-by: Steve Beattie <steve.beattie@canonical.com>