kernel-abi-check: convert to use usn_lib.USNdb
The kernel-abi-check script needs to determine what is the highest
version a given kernel/release has had a USN published for, in order to
determine if there has been a kernel published that has not had a USN
published for it yet.

Instead of exec'ing a separate process that has to re-open the USN db
each time to look up the information, use the (now friendlier)
usn_lib.USNdb class directly, to increase performance of the script.

Signed-off-by: Steve Beattie <steve.beattie@canonical.com>