sis-generate-usn: include source package name for all-binaries section
For some USN types, in particular for kernels, we end up removing
some of the source packages, but we want to retain the association
with the source for linking and other reasons, so include them as
part of the all-binaries section.

This will allow them to be included as part
of the USN website, and will help address
https://github.com/canonical/ubuntu-advantage-client/issues/1465

This commit depends on the merge request in
https://code.launchpad.net/~sbeattie/usn-tool/+git/usn-tool/+merge/403922
having landed.

Signed-off-by: Steve Beattie <steve.beattie@canonical.com>
Bug: https://github.com/canonical/ubuntu-advantage-client/issues/1465
MR: https://code.launchpad.net/~sbeattie/ubuntu-cve-tracker/+git/ubuntu-cve-tracker/+merge/403924