publish-usn-to-website-api.py: use source info from allbinaries section
With sis-generate-usn.py set to add source name references into the
allbinaries section, adjust the usn publishing script to generate
launchpad source links and version URLs based on this information.
Emit a warning if the referred source name is not found in the sources
section for that release.

Kernel USNs being the extra special things that they are have
binaries listed where the source package has been dropped. In that
case, the versions of the binaries match the version of the source,
so hackishly use the binary version. Additionally, inform the use of
this grotesque hack.

Signed-off-by: Steve Beattie <steve.beattie@canonical.com>
MR: https://code.launchpad.net/~sbeattie/ubuntu-cve-tracker/+git/ubuntu-cve-tracker/+merge/403924
Bug: https://github.com/canonical/ubuntu-advantage-client/issues/1465