sis-generate-usn: don't apply arg filter for allbinaries
The kernel publication process script prepare-kernel-usn.py passes
a command line argument to sis-generate-usn that attempts to filter
the set of applicable binaries for inclusion in a USN to just the
specific linux-image* binary and meta packages, eliding things like
the associated headers. However, these should be included in the
allbinaries section for tools like 'ua fix' to be able to find.
So drop the filter when emitting the allbinaries entries.

Signed-off-by: Steve Beattie <steve.beattie@canonical.com>