kernel/CVE-2021-0606: add break-fix commits
The fix commit is specific to v4.14.y. It is fixing a bad backport of a
commit that was present in v4.15.0. Because of that, the break commit is
the backport commit, not the original one. Otherwise, our kernels would be
shown as broken, with no fix available, because no fix is necessary.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>