scripts/check-syntax-fixup: Add script to fix check-syntax errors
This is intended to read the output of check-syntax and automatically try
and remediate the issues. For now it handles the cases of a CVE file
missing entries for supported releases, as well as CVE files containing
entries which should not be there too. Since check-syntax outputs to
stderr, it should be run as follows to read directly from the output:

./scripts/check-syntax 2>&1 | ./scripts/check-syntax-fixup

Or you can redirect check-syntax output to a file and read from that:

./scripts/check-syntax 2>&1 1> /tmp/check-syntax-output
./scripts/check-syntax-fixup /tmp/check-syntax-output