scripts/html_export.py: Output patches div even if no actual patches
This allows the generated HTML to more closely represent the underlying CVE
file - ie if a CVE file has

Patches_foo:

But with no entries, we should still generate the patches div/table but we
leave it empty in this case. But if it has no Patches_foo: entry at all
then we don't generate anything.