publish-cves-to-website-api: many CVEs do not have a Mitigation section
The publish-cves-to-website-api.py script was failing due to key errors
looking up a Mitigation field that doesn't exist for many retired CVEs.
Use the default value of an empty string instead; this is what
cve_lib.load_cve() puts in the field when the Mitigation: field in the
entry is empty.

Signed-off-by: Steve Beattie <steve.beattie@canonical.com>