kernel/CVE-2022-23222: replace break-fix commits
As a break commit, use the commit that introduces PTR_TO_MEM, which is the
commit that introduced ringbuf maps.

As a fix commit, though mainline requires a patch series, use the point in
that series that really fixes the issue in question, that is, the test in
adjust_ptr_min_max_vals. As for the older kernels, use the title for the
commit that was introduced in upstream stable series.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>