kernel/CVE-2018-25020: add break commit
Constant blinding is one of the things that makes the JITed programs large
enough to trigger the issue. Use the commit that enables constant blinding
on x86 JIT as the break commit. On our 4.4 kernels, we carry the constant
blind code, but do not use it on any JIT. This was all introduced on 4.7.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>