CVE-2020-13627: move to not-for-us
Not sure how I messed up the initial triage; but this CVE affects the
web interface for centreon apps, which is not included as part of the
centreon-engine source (where I had intended the CVE to be targeted
against). There is a debian intent-to-package (itp) bug to package up
the web interface, including that reference.

Signed-off-by: Steve Beattie <steve.beattie@canonical.com>