pkg_status: use bash so we can "set -o pipefail"
The ubuntu-table script is invoked by pkg_status in a pipeline, and bugs
or problems can cause it to abort. When that happens, without pipefail
set, the errors are eaten silently and no output is returned. With
pipefail set, at least pkg_status will also return an error, though that
won't be noticed in most people's shells.

Signed-off-by: Steve Beattie <steve.beattie@canonical.com>