kernel/CVE-2022-1116: add break alternative commit ID
Breaking commit cac68d12c531 ("io_uring: grab ->fs as part of async offload")
was specific to 5.4.y and did not have an upstream commit (it was two upstream
commits in one, in fact).

When applying such commits to our 5.4 tree, the 5.4.y commit ID is not present
in the commit message, so the triage bot does not detect the commit presence.
Use the commit ID used in our 5.4 tree in addition to the original commit ID as
a break commit.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>