oval_lib: Fix 2 issues in OVAL logic regarding kernel vulnerabilities
1. We were incorrectly checking if the patched kernel is greater than
or equal to running kernel. We should only check if patched kernel is
greater than running kernel, that way oscap will return True only when
fix is missing. This affected both USN and CVE OVAL.

2. On CVE OVAL we were incorrectly re-using elements such as  objects,
variables and tests for kernel version comparison, causing the comparison
to be done with a version that was not the one fixed. USN OVAL doesn't
suffer from it as is doesn't have re-use of elements.