report-todo-mir: reduce redundant bug queries
Given a task t, asking for t.bug.FIELD results in a round trip to
launchpad to query the bug again. Reduce the times we do this by caching
a reference to the bug itself. The biggest change here was to lift the
allocation of a MIRTask object (formerly MIRBug) out of the print_bugs()
function and up to where we first are given a collection of bug tasks.

This reduces the number of times a bug is queried from launchpad from 4
for each task down to one (plus one additional query to get the cves if
any associated with the bug). Not a huge win, but still an improvement.

Signed-off-by: Steve Beattie <steve.beattie@canonical.com>