report-todo-mir: convert bug dict to a class
Convert the bugs dictionary to a full blown class; this lets us hide
some of the details about how these things are stored and sorted in the
class. It is also useful for future parallelization work, as adding
tasks will need to be synchronized across whatever parallel querying
tasks occur.

Signed-off-by: Steve Beattie <steve.beattie@canonical.com>