kernel: add fix commits for MMIO stale data mitigations
Fixes were applied to Ubuntu kernels using upstream commit IDs. There were
multiple patches required to mitigate the issues, but the most important ones
are the ones listed here. Each CVE got a different commit, in an effort to
track those 3 different commits are applied, but also to match vulnerability
and mitigation.

There is no break commit as these are really hardware vulnerabilities, not
Linux ones. Though given the limitation of the mitigation, it could be argued
that VFIO introduced the vulnerability.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>