scripts/boilerplate-to-json.py: Script to convert boilerplates to json
The idea is to use this once to generate a meta_lists/package-db.json which
will contain the details captured from the existing boilerplate files, as
well as the meta_lists/package_info_overrides.json so we have one place to
store this kind of information. Subsequent commits will convert the various
tools to use this file instead of the boilerplates.

Signed-off-by: Alex Murray <alex.murray@canonical.com>