check-cves: Suggest the preferred package name instead of an alias
meta_lists/package-db.json supports the concept of an alias list for a given
package - and since cve_lib adds these as alternative keys to lookup a package
in cve_lib.package_db we can then match on these as well as the preferred
package name. But if we then show this alias name to a human when doing CVE
triage this will likely be confusing since say in the case of libvncclient, this
is a library embedded in libvncserver - so we can add an alias of libvncclient
to the libvncserver entry in package-db.json - but then we want the suggested
package name to be libvncserver - so we need to know this is the preferred name
for CVE triage. As such, we add a new name field to the loaded package_db
dictionary in cve_lib that can be used as this preferred name, and then look
this back up before providing the list of suggestions to the user during CVE
triage in check-cves. As such, a CVE where the text suggest libvncclient will
the be offered to add a CVE against libvncserver as expected.

Signed-off-by: Alex Murray <alex.murray@canonical.com>