scripts/cve_lib.py: Remove FileNotFoundError in load_package_db()
This file is checked into UCT git now so we don't want to catch this error and
pass it - instead don't catch it at all and let it bubble up since if this fails
then we have bigger problems on our hands as this file should now always exist.

This then conveniently fixes the use of cve_lib.py on python2 where this
exception is not defined.

Signed-off-by: Alex Murray <alex.murray@canonical.com>