Revive boilerplates/package_info_overrides.json to replace package-db.json
However, rather than storing the boilerplate files in the active subdirectory,
store them in their own separate directory. Also we parse these on-the-fly in
cve_lib to create the package-db structure rather than having this simply stored
as a big json file on disk. This should allow the ease of use of editing that
the boilerplate files provided coupled with the benefits of the smarter CVE
creation etc that package-db provides.

Signed-off-by: Alex Murray <alex.murray@canonical.com>