scripts/active_edit: Fix handling of non-canonical boilerplates
Don't discard package names too early otherwise we will miss adding package
entries for non-canonical boilerplates. cve_lib treats real files under
boilerplates/ as the canonical package name for a given package - and so the
chromium would be listed as the name for the chromium and chromium-browser
entries in package_db - as such, if a user tried to add a CVE with the package
chromium-browser, this name would get normalised to 'chromium' but then we would
discard that name as there is no 'chromium' package in Ubuntu. So instead we
should only discard the package name at the point we go to add it from the pkgs
entry rather than earlier.

Signed-off-by: Alex Murray <alex.murray@canonical.com>