kernel/CVE-2022-3625: find break-fix commits
This patch fixed an issue caused by lack of sound checking for
`devlink_param_get` and `devlink_param_set` functions, and these two functions
are introduced in the commit 45f05def5c44c806f094709f1c9b03dcecdd54f0.

[cascardo:

The reported vulnerability is on mlxsw driver, and the Fixes: commit is the
one introducing that vulnerability, which would be
98bbf70c1c41fb9547c3a18c0f1b96f6ebb8eb1d.

However, we are opting to use this other one in case other drivers or the
same driver decide to rely on that mitigation and introduce new features
with the same kind of vulnerability.

]

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>