kernel/CVE-2022-4378: triage
Prioritize as a High, as user/network namespaces may be used to trigger a
stack overflow.

The break commit here is used because before it, the buffers were limited
to 'left' length and NULL-terminated. After that commit, buffers started
being copied on proc_sys_call_handler with the full length given by
userspace, instead of only up to PAGE_SIZE - 1.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>