publish-cves-to-website-api.py: don't push impact field when no CVSS score is present
The commit d863026cc2 ("scripts/publish-cves-to-website-api.py: Add
impact field") added support for sending the detailed CVSS vector to the
CVE website, but did so in a way that attempts to set the field to None
when no CVSS score has been assigned to the CVE, which the CVE website
rejects.

Fix this by only populating the "impact" field in the submitted JSON
when we have an actual vector to submit.

Fixes: d863026cc2 ("scripts/publish-cves-to-website-api.py: Add impact field")
Signed-off-by: Steve Beattie <steve.beattie@canonical.com>
Acked-by: Paulo Flabiano Smorigo <pfsmorigo@canonical.com>