CVE-2022-4696: Add break-fix line
[updated by sbeattie]

The fixing commit 75454b4bbfc7 ("io_uring: add missing item types
for splice request") was only applied to the 5.10-stable branch;
the 5.12 era commit 44526bedc2ff ("io_uring: remove any grabbing of
context") got rid of the work_flags field entirely. So use the later
commit as the fixing commit and include the stable branch commit as
an alternative fix in the cve and in the "special cases" definition
file active/10autotriage.linux.

Signed-off-by: John Cabaj <john.cabaj@canonical.com>
Message-Id: <20230125205319.556150-1-john.cabaj@canonical.com>
Signed-off-by: Steve Beattie <steve.beattie@canonical.com>