source_map.py: adjust alias membership test to prevent traceback
In python, `if foo["bar"]:` will throw a traceback by default if there's
no "bar" key in foo. The test should be `if "bar" in foo:` to skip if
the key is not in the dict. This manifests in alias code in
source_map.py like so:

  Traceback (most recent call last):
    File "/home/steve/git/ubuntu-cve-tracker/./scripts/check-cves", line 68, in <module>
      source = source_map.load()
    File "/home/steve/git/ubuntu-cve-tracker/scripts/source_map.py", line 273, in load
      subproject_lists = load_subprojects_lists(releases=releases)
    File "/home/steve/git/ubuntu-cve-tracker/scripts/source_map.py", line 479, in load_subprojects_lists
      if map[rel][pkg]:
  KeyError: 'python-cryptography'

Fix this by converting the `if map[rel][pkg]:` test to `if pkg in map[rel]:`

Signed-off-by: Steve Beattie <steve.beattie@canonical.com>