kernel/CVE-2022-3903: fix commit came through two separate merges
The commit 608e58a0f4617977178131f5f68a3fce1d3f5316 ("media:
mceusb: Use new usb_control_msg_*() routines") was part of
two separate merges of the media tree, in v6.0 and v6.1 (as
41fd1cb6151439b205ac7611883d85ae14250172). We'll treat the earlier
commit (608e58a0f461) as the fixing commit, but the kernel triage
bot may only be finding the latter commit, so alias it to the former.

To see the separate merges, do

   git log --show-pulls drivers/media/rc/mceusb.c

on both the linux-6.0.y and linux-6.1.y stable series branches;
you should see:

  linux-6.0.y: fd59585c420d ("Merge tag 'usb-6.0-rc4' of git://git.kernel.org/pub/scm/linux/kernel/git/gregkh/usb")
  linux-6.1.y: 3272eb1ace32 ("Merge tag 'media/v6.1-2' of git://git.kernel.org/pub/scm/linux/kernel/git/mchehab/linux-media")

with the latter containing the following explanatory test:

   Pull missed media updates from Mauro Carvalho Chehab:
     "It seems I screwed-up my previous pull request: it ends up that only
      half of the media patches that were in linux-next got merged in -rc1.

      The script which creates the signed tags silently failed due to
      5.19->6.0 so it ended generating a tag with incomplete stuff.

Thanks to Yuxuan for doing the initial research and identifying
both commits.

Signed-off-by: Yuxuan Luo <yuxuan.luo@canonical.com>
Signed-off-by: Steve Beattie <steve.beattie@canonical.com>