kernel/CVE-2023-0240: fixup break-fix
The commit used as fix is actually the break commit. That can be inferred
from the vulnerability description, which mentions io_grab_identity being
called from io_prep_async_work. And that is the code doing get_cred.

As for the real fix, it is the one removing all of that, as kthreads have
been replace by proper sibling threads.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>