kernel/CVE-2023-22998: adapt break commit
The original break commit came from the Fixes: line of the fix commit, but it
is not correct. That commit only moves code around. The proposed break commit
is the one where virtio gpu is converted to GEM and start using the helper that
despite its documentation stating that it returns NULL on failure, it actually
returns an error pointer.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>