kernel/CVE-2022-20132: break-fix triage
c500c9714011edab021591340042787722db9cf0 is actually a dependency for two of
the three remaining commits. The other one is bigben specific and its break
commit is the introduction of this driver.

For wacom driver, the break commit is the first one that adds non-USB Wacom
support.

As for the last one, as it touches many different drivers, the break commit
chosen is the one that introduces the first non-USB low-level HID driver
(Bluetooth).

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>