kernel/CVE-2023-1249: remove one break commit
2aa362c49c314a98fb9aebbd7760a461667bac05 ("coredump: extend core dump note
section to contain file names of mapped files") is from 2012. And the fix
commit doesn't make sense without the second break commit from 2020.

So, unless we decide we need this newer break commit in order to fix this
older one, which will lead to us backporting both commits, let's leave this
very old commit out of the way for triage.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>