kernel/CVE-2022-0168: change break commits
The first fix actually requires the vars variable and a fix for a UAF/double
free when using it, which is not present on 5.4 kernels.

The second fix does not require the change from kmalloc to kmemdup_user, but
has two paths on it, and each one of them require a different hunk, so have two
different break commits.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>