source_map: load esm-infra the same way as esm-apps
Right now, we are loading the packages from ESM Infra from
the mirrors of the main release and then adding any extra
packages from supported.txt.

We cannot remove any packages by using supported.txt this way.
This commit changes this behaviour to load ESM Infra the same way
as ESM Apps and only rely on the information from this file as the
packages existing in the release.