scripts/check-syntax: complain if Priority reason is XXX-Reason-XXX
active_edit will put this as the placeholder when specifying a CVE priority
which requires a reason - in which case, cve_lib and check-syntax see this as a
reason and so never complain about one being missing. But we should still check
for this placeholder and make sure it gets updated to something real, so do this
in check-syntax.

Signed-off-by: Alex Murray <alex.murray@canonical.com>