kernel/CVE-2023-21400: temporary break-fix commits
We know this does not affect any kernels with no io_uring support, so use
the commit introducing it as lower bound. The right break commit still need
to be determined in order to mark 5.4 kernels as not affected.

Use the upstream stable commit as the fix commit such that 5.15 and 5.17
kernels are marked as fixed. The upstream fix commit still need to be
identified and used as an alternative fix.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>