kernel/CVE-2023-{2640,32629}: add alternative local fix
Given the original SAUCE patch was removed instead of reverted in later
releases, we need to use the alternative SAUCE that was introduced in its
place as a fix for these vulnerabilities.

The reason we cannot use the original SAUCE as a break commit is because it
was not reverted on earlier releases that did not contain the other commits
which lead to the vulnerability due to their interactions with the SAUCE.
Those are used as break commits, which allow us to determine the releases
that really needed the revert.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>