kernel-abi-check: add --use-lp-bot-account option to use alt lp creds file
Because we try to use reduced privilege bot accounts in shared vms to
access lp for automation purposes, we can sometimes have the same shared
unix user accessing launchpad through different lp ids. To partially
avoid this for the kernel-abi-check script, add an option that lets us
do that for this script, at least with the lpl_common.connect() API as
it currently stands.

Signed-off-by: Steve Beattie <steve.beattie@canonical.com>