cve_lib.py: improve subproject error reporting
THe commit b4db9b9938c ("Add external subprojects metadata validation")
unfortunately is too strict for some of the internal subprojects we have
configured, and breaks some of our processes. Unfortunately, the way
this was being reported resulted in the detailed message only showing up
in stdout, and a valueless ValueError being raised.

This commit adds a strict argument to load_external_subprojects() that
currently defaults to True (which doesn't solve the process breakage
issue) and then based on the strict value either raises the
ValueError with the detailed error message or prints it to stderr
and continues.

A temporary fix to set strict to default to False will be in a followup
commit, so that it can easily be reverted in a future commit.

Signed-off-by: Steve Beattie <steve.beattie@canonical.com>