publish-cves-to-website-api.py: reduce amount emitted on failure
When failures to push to the cve website api occur, the emitted json
output can be voluminous, in particular for kernel CVEs. reduce the
output and also emit the output to stderr, not stdout. *Ideally* this
would be recorded both in the host's logfile and reported by email, but
there's some madness there that needs tp be sorted out as well.

Signed-off-by: Steve Beattie <steve.beattie@canonical.com>