kernel/CVE-2024-0340: add break commit and adjust priority
The break commit should be 670ae9caaca4 (“vhost: fix info leak due to
uninitialized memory”) and is likely a follow-up "fix" for
CVE-2018-1118. Since the leaking issue should be solved by the above
commit and this fix commit only zeros memory that was not considered
leaking in CVE-2018-1118, adjusting the priority to low.

Signed-off-by: Yuxuan Luo <yuxuan.luo@canonical.com>
Message-Id: <ZaoWHNgux3B9PB2B@nxnw.org>
Signed-off-by: Steve Beattie <steve.beattie@canonical.com>