kernel/CVE-2024-0193: change break commit
The fix for CVE-2024-0193, 7315dc1e122c ("netfilter: nf_tables: skip set
commit for deleted/destroyed sets") has a Fixes: line pointing out to
5f68718b34a5 ("netfilter: nf_tables: GC transaction API to avoid race with
control plane").

However, 7315dc1e122c fixes a function introduced by 212ed75dc5fb
("netfilter: nf_tables: integrate pipapo into commit protocol"), which is
really what allows CVE-2024-0193 to be exploited.

If this latter commit is necessary to prevent other forms of attacks, we
cannot find a CVE assigned to it, and it has only been backported upstream
up to 5.10, not 5.4, which is the latest kernel missing the fix for
CVE-2024-0193.

Therefore, use this latter commit as the break commit for CVE-2024-0193.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>