Fix active_edit no longer copying statuses from boilerplates (LP: #2052658)
scripts/cve_lib.py: don't drop the package when loading the boilerplates
as we need to have the statuses later on.
scripts/active_edit: check for duplicate by comparing against p, not
against parent. I believe this is a typo as this code is only
reachable when parent is None.